# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Artist API
  version: 1.0.0
  description: API de gestion des artistes (acteurs / réalisateurs)

servers:
  - url: http://localhost:8081

tags:
  - name: artists
    description: Gestion des artistes

paths:
  /api/artists:
    get:
      tags: [artists]
      summary: Liste des artistes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistDto"
    post:
      tags: [artists]
      summary: Créer un artiste
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistDto"
      responses:
        "200":
          description: OK (créé)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistDto"

  /api/artists/{id}:
    get:
      tags: [artists]
      summary: Détail artiste
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistDto"
        "404":
          description: Non trouvé
    put:
      tags: [artists]
      summary: Modifier un artiste
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistDto"
      responses:
        "200":
          description: OK (modifié)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistDto"
    delete:
      tags: [artists]
      summary: Supprimer un artiste
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean

components:
  schemas:
    ArtistDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        prenom:
          type: string
        nom:
          type: string
        age:
          type: integer
          nullable: true
        genre:
          type: string
          nullable: true
        estActeur:
          type: boolean
          nullable: true
        estRealisateur:
          type: boolean
          nullable: true
        bio:
          type: string
          nullable: true
      required:
        - prenom
        - nom